refactor: clean and optimize the code
